+++ examples/data-sources/infrapilot_license_check/data-source.tf
@@ -1,6 +1,6 @@
 // Copyright (c) 2025 InfraPilot, LLC
 // SPDX-License-Identifier: MPL-2.0
 
-provider "infrapilot" {
-  token = "dummy-token"
+data "infrapilot_license_check" "validation" {
+  module = "example"
 }